MDW-1645: allowing to asynchronously inject the mapbox DEV JS. Allowing to set zoom as double.

+++ src/main/java/com/tomtom/gwt/mapbox/gl/client/MapboxMap.java
@@ -162,7 +162,7 @@ public class MapboxMap extends AbstractEvented {
         return (int) (Math.round(getZoom()));
     }
 
-    public native MapboxMap setZoom(int zoom);
+    public native MapboxMap setZoom(double zoom);
 
     public native LngLat getCenter();
